Disabled build of tiscamerasrc
The tiscamerasrc module has not been updated to correctly use
current tis modules.
The existing version is more of a "proof of concept" than a functioning
module. USB3 support is missing, as well as support for most USB2
variations. Until a better version is built this will not be built.