Fix shutdown issues for gige-daemon
gige-daemon suffered from issues when being run by Systemd.
The daemon did not shut down correctly when run by systemd resulting in
a 90 second waiting period.
The shared memory between processes has to be readable/writeable for
all.

The flag --no-fork was introduced to allow running of the daemon in the
foreground, thereby enabling the systemd run-type "simple".

The main thread now also uses sigsuspend to wait for signals in a more
dedicated way, thus preventing interupts that caused the systemd problems.