af: Work on read only buffers
The autofocus algorithm should not need to have a writable buffer so it
should be mapped read only. If the buffer needs to be writable, the
implementation needs to make sure to handle gst_buffer_make_writable
correctly (which it did not). Also this would impose a performance drop
if the buffer can not be made writable without copying.