indexing: Separate device indexing from DeviceIndex
With the DeviceIndex being static some users had problems when ending their
application. This commit makes the DeviceIndex a normal class that is treated
like any other class. The actual indexing functionality is moved to the Indexer
class. This class is shared between all DeviceIndex instances and will
automatically be deleted once all DeviceIndex instances cease to exist.
This results in the indexing thread ending before the static BackendLoader and
the dynamically loaded backends are removed.