uvc/udev: fix uvc extension loading for USB2 cameras
The udev rules for USB2 cameras compared the first byte of the USB
product ID to the speed attribute. This problem was introduced in commit
2adc718 ("uvc/udev: Fix extension unit handling for USB3 cameras") when
changing from speed comparison to USB product IDs.

This commit fixes that udev rule by checking the correct attribute
against the first byte of the USB product ID.

Fixes: 2adc718 ("uvc/udev: Fix extension unit handling for USB3 cameras")
Signed-off-by: Richard Leitner <richard.leitner@skidata.com>