cmake: Remove CMAKE_SOURCE_DIR and CMAKE_BINARY_DIR
To ensure tiscamera can be used as a submodule/subproject
the usage of CMAKE_SOURCE_DIR and CMAKE_BINARY_DIR has to be avoided.
These point to the parent project and not to the tiscamera root directories.
Instead the variable TCAM_SOURCE_DIR and TCAM_BINARY_DIR have been introduced.